hello-ms-dj/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
drf-spectacular==0.26.5
eureka==0.15
ifaddr==0.2.0
inflection==0.5.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
py-eureka-client==0.11.8
PyJWT==1.7.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.12.0
sqlparse==0.4.4
uritemplate==4.1.1
